exported table, loading and error components into separate files
