+++ .eslintrc.cjs
@@ -22,5 +22,6 @@ module.exports = {
   rules: {
     'react/react-in-jsx-scope': 'off',
     'import/prefer-default-export': 'off',
+    'react/require-default-props': 'off',
   },
 };
