+++ src/components/index.ts
@@ -1,0 +1,3 @@
+export { default as LoadingPlaceholder } from './LoadingPlaceholder';
+export { default as EmptyPlaceholder } from './EmptyPlaceholder';
+export { default as GithubReposTable } from './GithubReposTable';
